hj6j

+++ app.py
@@ -47,10 +47,10 @@ else:
 
 #brand_df.drop(['Unnamed: 0','Brand_Family_Desc','region_desc'],axis=1,inplace=True)
 
-brand_df['Advertisement_and_discount_cost']=brand_df['Advertisement_and_discount_cost'].astype('float64')
-brand_df['Transfer_Price_COGS']=brand_df['Transfer_Price_COGS'].astype('float64')
-brand_df['Distribution_cost_supply_chain']=brand_df['Distribution_cost_supply_chain'].astype('float64')
-brand_df['Distribution_cose_Sales_-_Region']=brand_df['Distribution_cose_Sales_-_Region'].astype('float64')
+#brand_df['Advertisement_and_discount_cost']=brand_df['Advertisement_and_discount_cost'].astype('float64')
+#brand_df['Transfer_Price_COGS']=brand_df['Transfer_Price_COGS'].astype('float64')
+#brand_df['Distribution_cost_supply_chain']=brand_df['Distribution_cost_supply_chain'].astype('float64')
+#brand_df['Distribution_cose_Sales_-_Region']=brand_df['Distribution_cose_Sales_-_Region'].astype('float64')
 
 brand_df['Advertisement_and_discount_cost']=brand_df['Advertisement_and_discount_cost'].fillna(brand_df['Advertisement_and_discount_cost'].median())
 brand_df['Transfer_Price_COGS']=brand_df['Transfer_Price_COGS'].fillna(brand_df['Transfer_Price_COGS'].median())
